+++ src/main/kotlin/com/aoc/intcode/amplifier/series/OutputAmplifier.kt
@@ -4,10 +4,10 @@ import com.aoc.intcode.IntCodeComputer
 import com.aoc.intcode.amplifier.Amplifier
 import com.aoc.intcode.amplifier.BaseAmplifier
 
-class OutputAmplifier(private val phaseSetting: Int) : BaseAmplifier(phaseSetting) {
-    private var outputSignal: Int = 0
+class OutputAmplifier(private val phaseSetting: Long) : BaseAmplifier(phaseSetting) {
+    private var outputSignal: Long = 0
 
-    override fun inputSignal(inputSignal: Int) {
+    override fun inputSignal(inputSignal: Long) {
         computer.getProgramMemory().systemInput(phaseSetting)
         computer.getProgramMemory().systemInput(inputSignal)
         computer.compute()
